src/listener/requirements.txt
telethon 
python-dotenv
pymongo